Move case-service onto loco-rs 1.0.1
Fourth crate on the family-wide loco-rs 1.0.1 migration; same recipe as
care-pathway-service (its closest sibling — both carry the audit
hash-chain and the IEC 62304 SOUP gate). Same mechanical fixes: raw
Statement calls to execute_raw across models/audit_logs.rs,
compliance/erasure.rs, and tests/requests/cases.rs; PkAuto's 64-bit
width reaching cases/audit_logs/merge_records plus the hash-chain's own
id-carrying types; DatabaseConnection::Disconnected replaced by an empty
MockDatabase; a useless .into() in event_outbox.rs.

No SOUP register change — loco-rs and sea-orm were already annotated.

Verified: cargo build --all-targets, fmt, clippy -D warnings, cargo test
(176 passed) clean; the full DB-gated suite (34 tests, unchanged count)
green against a freshly migrated Postgres 18.

+++ case/case-service-with-loco/src/compliance/audit_chain.rs
@@ -165,7 +165,7 @@ pub fn input_for<'a>(row: &'a audit_logs::Model, prev: Option<&'a str>) -> Chain
 #[derive(Debug, Clone, Serialize, PartialEq, Eq)]
 pub struct ChainBreak {
     /// The `audit_logs.id` at which the break was detected.
-    pub id: i32,
+    pub id: i64,
     /// `linkage` — the row does not point at its predecessor's hash (a row
     /// was inserted, deleted, or reordered); `content` — the row's stored
     /// hash does not match its own content (the row was edited).
@@ -389,14 +389,14 @@ mod tests {
     /// Build a chained row the way the writer does: hash over the row's own
     /// content plus `prev`.
     fn row(
-        id: i32,
+        id: i64,
         prev: Option<&str>,
         action: &str,
         snapshot: Option<serde_json::Value>,
     ) -> audit_logs::Model {
         let mut model = audit_logs::Model {
-            created_at: at(1_700_000_000_000_000 + i64::from(id)),
-            updated_at: at(1_700_000_000_000_000 + i64::from(id)),
+            created_at: at(1_700_000_000_000_000 + id),
+            updated_at: at(1_700_000_000_000_000 + id),
             id,
             entity_pid: Uuid::from_u128(u128::from(id.unsigned_abs())),
             action: action.to_string(),
@@ -421,7 +421,7 @@ mod tests {
     }
 
     /// Build a well-formed chain of `n` rows.
-    fn chain(n: i32) -> Vec<audit_logs::Model> {
+    fn chain(n: i64) -> Vec<audit_logs::Model> {
         let mut rows: Vec<audit_logs::Model> = Vec::new();
         for id in 1..=n {
             let prev = rows.last().and_then(|r| r.hash.clone());
